pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["exp_logger", "exp_logger.*"]

[project]
name = "exp_logger"
version = "0.0.0-dev0"
description = "A simple experiment logger."
dependencies = [
    "tensorboardX",
    "pyyaml"
]


readme = "README.md"
requires-python = ">=3.8"
authors = [
  {name = "Haitong Ma", email = "haitongma@g.harvard.edu" }
]
maintainers = [
  {name = "Haitong Ma", email = "haitongma@g.harvard.edu" }
]

[project.urls]
